requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.2
django-ratelimit==2.0.0
django-redis==4.11.0
django-rq==2.2.0
djangorestframework==3.11.0
idna==2.8
mysqlclient==1.4.6
pytz==2019.3
redis==3.4.1
requests==2.22.0
rq==1.2.2
sqlparse==0.3.0
urllib3==1.25.8
